[UPD] pinned projects
change the way I get my github pinned projects, and use an external library for my icons, also adding my email. For the future, I will need to refacto the main page, also changing the first secion of the site to add some infos or something that's cool to see. Without forgetting, if possible to add a second view of the project when clicking on it.